Initial reordering.
