+++ src/chromatix/functional/__init__.py
@@ -5,3 +5,4 @@ from .propagation import *
 from .samples import *
 from .sensors import *
 from .polarizers import *
+from .convenience import optical_fft
