+++ src/chromatix/ops/__init__.py
@@ -1,9 +1,5 @@
-from .ops import (
-    fourier_convolution,
-    pooling_downsample,
-    init_plane_resample,
-)
+from .ops import *
 from .noise import *
-from .windows import *
+from .resample import *
 from .filters import *
 from .quantization import *
